large disk support

+++ ldmutil/diskio.h
@@ -18,12 +18,12 @@ public:
 	~diskio(void);
 	void Open(const char* filename, bool readonly = true);
 	void Close(void);
-	static int GetSectorSize(void);
-	void SetPos(int sector);
-	int  GetPos(void);
-	void Write(const void* src, int nsect = 1, int pos = INT_MIN);
-	void Read(void* dest, int nsect = 1, int pos = INT_MIN);
-	int  GetSize(void);
+	static u64 GetSectorSize(void);
+	void SetPos(u64 sector);
+	u64  GetPos(void);
+	void Write(const void* src, int nsect = 1, u64 pos = INT_MIN);
+	void Read(void* dest, int nsect = 1, u64 pos = INT_MIN);
+	u64  GetSize(void);
 };
 
 }
